requirements.txt
langchain
openai==0.28.1
huggingface_hub
python-dotenv
streamlit